Add skeleton cli crate
