+++ crates/fe/src/main.rs
@@ -1,0 +1,7 @@
+use clap::Parser;
+use driver::Options;
+
+fn main() {
+    let opts = Options::parse();
+    driver::run(&opts);
+}
