+++ crates/uitest/tests/ty.rs
@@ -1,5 +1,4 @@
-use std::path::Path;
-
+use camino::Utf8Path;
 use dir_test::{dir_test, Fixture};
 use driver::DriverDataBase;
 use test_utils::snap_test;
@@ -10,7 +9,7 @@ use test_utils::snap_test;
 )]
 fn run_ty_def(fixture: Fixture<&str>) {
     let mut db = DriverDataBase::default();
-    let path = Path::new(fixture.path());
+    let path = Utf8Path::new(fixture.path());
 
     let (ingot, file) = db.standalone(path, fixture.content());
     let top_mod = db.top_mod(ingot, file);
@@ -26,7 +25,7 @@ fn run_ty_def(fixture: Fixture<&str>) {
 )]
 fn run_const_ty(fixture: Fixture<&str>) {
     let mut db = DriverDataBase::default();
-    let path = Path::new(fixture.path());
+    let path = Utf8Path::new(fixture.path());
 
     let (ingot, file) = db.standalone(path, fixture.content());
     let top_mod = db.top_mod(ingot, file);
@@ -42,7 +41,7 @@ fn run_const_ty(fixture: Fixture<&str>) {
 )]
 fn run_trait_bound(fixture: Fixture<&str>) {
     let mut db = DriverDataBase::default();
-    let path = Path::new(fixture.path());
+    let path = Utf8Path::new(fixture.path());
 
     let (ingot, file) = db.standalone(path, fixture.content());
     let top_mod = db.top_mod(ingot, file);
@@ -58,7 +57,7 @@ fn run_trait_bound(fixture: Fixture<&str>) {
 )]
 fn run_trait_impl(fixture: Fixture<&str>) {
     let mut db = DriverDataBase::default();
-    let path = Path::new(fixture.path());
+    let path = Utf8Path::new(fixture.path());
 
     let (ingot, file) = db.standalone(path, fixture.content());
     let top_mod = db.top_mod(ingot, file);
@@ -88,7 +87,7 @@ mod wasm {
         )]
         fn run_ty_def(fixture: Fixture<&str>) {
             let mut db = DriverDataBase::default();
-            let path = Path::new(fixture.path());
+            let path = Utf8Path::new(fixture.path());
 
             let (ingot, file) = db.standalone(path, fixture.content());
             let top_mod = db.top_mod(ingot, file);
@@ -109,7 +108,7 @@ mod wasm {
         )]
         fn run_const_ty(fixture: Fixture<&str>) {
             let mut db = DriverDataBase::default();
-            let path = Path::new(fixture.path());
+            let path = Utf8Path::new(fixture.path());
 
             let (ingot, file) = db.standalone(path, fixture.content());
             let top_mod = db.top_mod(ingot, file);
@@ -130,7 +129,7 @@ mod wasm {
         )]
         fn run_trait_bound(fixture: Fixture<&str>) {
             let mut db = DriverDataBase::default();
-            let path = Path::new(fixture.path());
+            let path = Utf8Path::new(fixture.path());
 
             let (ingot, file) = db.standalone(path, fixture.content());
             let top_mod = db.top_mod(ingot, file);
@@ -151,7 +150,7 @@ mod wasm {
         )]
         fn run_trait_impl(fixture: Fixture<&str>) {
             let mut db = DriverDataBase::default();
-            let path = Path::new(fixture.path());
+            let path = Utf8Path::new(fixture.path());
 
             let (ingot, file) = db.standalone(path, fixture.content());
             let top_mod = db.top_mod(ingot, file);
