+++ crates/uitest/tests/ty_check.rs
@@ -1,5 +1,4 @@
-use std::path::Path;
-
+use camino::Utf8Path;
 use dir_test::{dir_test, Fixture};
 use driver::DriverDataBase;
 use test_utils::snap_test;
@@ -10,7 +9,7 @@ use test_utils::snap_test;
 )]
 fn run_ty_check(fixture: Fixture<&str>) {
     let mut db = DriverDataBase::default();
-    let path = Path::new(fixture.path());
+    let path = Utf8Path::new(fixture.path());
 
     let (ingot, file) = db.standalone(path, fixture.content());
     let top_mod = db.top_mod(ingot, file);
@@ -36,7 +35,7 @@ mod wasm {
     )]
     fn run_ty_check(fixture: Fixture<&str>) {
         let mut db = DriverDataBase::default();
-        let path = Path::new(fixture.path());
+        let path = Utf8Path::new(fixture.path());
 
         let (ingot, file) = db.standalone(path, fixture.content());
         let top_mod = db.top_mod(ingot, file);
